Introduce IMX636 driver for 5.15 kernel
At this stage, the driver only registers itself, probes the sensor and
powers it when streaming (using PowerManagement framework from the kernel,
thus it should be feasible to play around with it.

The output format in hard-coded to MEDIA_BUS_FMT_PSEE_EVT3, which does not
exist in mainline kernel, moreover it won't output any data.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>