Implement CSI-2 reconfiguration
The operation requires to update some clock values in the chip while not
using it, and then update the CSI link timings according to its new freq.

Some register names are not really descriptive, but please refrain from
renaming then: that's their name in the actual RTL, and it's easier to
get support from the sensor design team that way.

The values above are pre-computed for the 20MHz input clock. Supporting
other values requires significant consideration.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>