Declare Evt2.1 Legacy for vanilla kernels
Evt2.1 legacy is just Evt2.1 with the 2 32-bit words switched, but the
userland needs this information.

As this worked started the IMX636, which has this weird endianness, the
medias bus code is kept for the middle-endian flavor.

The next patch on the driver will fix the advertised format, you may
get weird data if you build precisely on this commit. Sorry if your
bissect ends up here.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>