Fully start the sensor in pm_ops but not in probe
This work was motivated by an attempt to clearly split the
initialization and the start of the streaming of the sensor, to
hack some stuff in the middle.
This avoids being able to change the reported output format
while streaming (while not changing it until the next stream).
This will also help adding sensor controls in future developments.
There is unfortunately a documented race condition in this code,
that may happen if the format and stream status are managed by
two different threads/processes.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>