imx636: Allow TRY format change while streaming
Try is purely prospective, it is not an issue for the actual streaming.
The reason for this change is that, on device open, the
imx636_set_pad_format method is used to set the file handler. Without
this patch, it is not possible to open the v4l2-subdev while streaming.
It was not caught before because the media pipeline still works fine,
and in Prophesee Metavision, the device is opened before streaming, and
is kept opened while the application is running.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>