imx636: add V4L2 controls for analog bias
The biases are set according to IMX636-AAMR-C Application Note.
The bgen struct is fixed, because it was 33 bits large. Fortunately,
it was an extra MSB bit of no consequence on little-endian machines.
The bias_diff is tuning is now done by the v4l2_ctrl_handler_setup
when the sensor is powered-up.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>