imx636: add ROI control through selection (crop) API
The sensor allows to disable lines and columns, which allows more
complex patterns than a simple crop, but this covers the most common
use cases with an existing API.

A specific behavior of this driver is that the output format does not
change when a crop is applied. No events are generated outside of the
crop, but the events are not translated, a event generated in the top
left-hand corner of the crop is not marked as (0,0). The benefit of
that implementation is that crop can be set to full any size, up to
full resolution, during streaming.

As the video is a stream of sparse events, the output remaining at
full resolution creates no computational overhead, however, for
applications generating some kind of integration over the whole sensor
(histograms, dense optical flows, ...) this may cause extra memory
allocation, for areas where no event will be generated. This may be
addressed in a future work, for instance using OOR and composition API

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>