imx636: add tracing on CCI transfers
Just add debug traces in imx636_read_reg and imx636_write_reg.
It proved useful in many debug cases and it is annoying to rebuild the
kernel module to get it, this allow to get traces using dynamic debug.

When the kernel is built with it, write traces are obtained with
echo 'file imx636.c func imx636_write_reg +p'>/sys/kernel/debug/dynamic_debug/control

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>