imx636: fix CSI-2 frequency vs bitrate
The listed frequencies were actually bitrate. This may cause trouble
if the receiver uses this information to setup a PLL or similar HW
that needs to lock on CSI-2 clock.
While fixing the implementation, this patch can cause issue on system
whose device-tree listed the supported bitrate instead of the frequency.
One such system is the Kria KV260, where, fortunately, the clock was set
to 600Mb/s, and the hardware supports a 600MHz clock.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>