imx636: Tell CSI-2 information instead of sensor geometry in media-ctl
This will break some userspace, which won't find the sensor geometry
in the V4L2 device anymore. It may however use the crop bounds to get
the same information.

On the other hand, many camera acquisition pipeline need the line
length, and use the height to allocate the memory buffers. In this
patch, the width is set to the actual packet size used in the sensor,
while the height gives buffers large enough for the maximum throughput
of the internal pipeline of the sensor, which is higher than the
maximum output, resulting in memory over-reservation, but it would be
tricky to fine-tune as the framing works with the data time, and not
directly the CSI-2 clock.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>