genx320: implement LINK_FREQ control
This control is used by some receivers, probably to configure some PLL.
This implementation won't work if more frequencies get supported, but
it will allow to support more receivers in the meantime.

One issue is that the previous code referenced a bitrate instead of a
line frequency, and an erroneous one. Updating the driver will break
with previously working devicetrees.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>