imx636: Add a control to set End of Frame markers
The sensor is able to add 4 or 8 bytes of arbitrary data
at the end of the CSI-2 frames. This comes handy with
receivers whose DMA don't advertise the actual size of
the transfer. The default value is set to 0xE019E019,
which is a valid OTHERS event in all encodings.
In Evt3.0, the subtype is 0x019, while in Evt2.0 and 2.1
it is 0xE019, none of which are normally generated in
the sensor.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>