imx636: update how bias CIDs are defined
V4L2_CID_USER_BASE is 0x00980900, this means that, in bias base, the
8 in 0x1800 is useless. This commit mimics the implementations of other
drivers with custom controls, using a straight addition instead of a
logical OR, avoiding unexpected collisions with different masks. The
written offset is updated to keep the resulting CID, as the is no need
to change it.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>