imx636: add an option in the dt binding for fixed packet size
This is actually a feature of the receiver: some require a fixed packet
size within a frame, and usually process data frame per frame, making
packet latency irrelevant. As I write this commit, there is no interface
to tell that the receiver cannot handle variable size. Instead of patching
every receiver, it feels easier to write in the sensor binding that it
is used in such a system.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>