imx636: silence EPROBE_DEFER errors on probe
When the sensor is probed before its clock generator, it is expected
to get EPROBE_DEFER when requesting its clock, and to return the same
code to the kernel. The code already did that, but logged errors in
the kernel messages, as if something when wrong. With this commit,
the error is silently taken into account, and the kernel may retry
later.

Signed-off-by: Christian Dreher <cdreher@prophesee.ai>